fix(desktop): harden vault creation inputs
Reject obviously invalid cross-platform vault names before touching the filesystem and keep the desktop bridge contract in one place to reduce drift across main, preload, and creation code.
